cleaning up traits and renaming files
