+++ erdos/src/dataflow/operators/mod.rs
@@ -6,12 +6,12 @@ pub mod ros;
 
 // Private submodules
 // mod join_operator;
-mod filter_operator;
-mod map_operator;
-mod split_operator;
+mod filter;
+mod map;
+mod split;
 
 // Public exports
 // pub use crate::dataflow::operators::join_operator::JoinOperator;
-pub use filter_operator::FilterOperator;
-pub use map_operator::MapOperator;
-pub use split_operator::SplitOperator;
+pub use filter::{Filter, FilterOperator};
+pub use map::{Map, MapOperator};
+pub use split::{Split, SplitOperator};
